Backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2==2.9.11
asyncpg==0.30.0
alembic==1.13.1

# Authentication & Security
PyJWT==2.8.0
passlib==1.7.4
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.3.0
dnspython==2.8.0

# Date & Time
python-dateutil==2.8.2
